Wired in first service to use real stream data
